+++ frontend/components/JobCard.tsx
